requirements.txt
gTTS==2.5.0
moviepy==1.0.3
mutagen==1.47.0
numpy==1.24.3
Pillow==10.1.0
Pillow==10.1.0
pydub==0.25.1
whisperx==3.1.1
